Image Issue and Small Updates
The chalkboard background image was missing and I added a .chalkboard image section to provide a link to what will be the image we'll use for the background of the chalkboard search section.

Besides that I removed the .product p font and color since the product grid will not display the costs of the tempaltes.

This section will serve as a page link call to action button that will transition the user to that particular area of the store. For example the titles of the product grid portray what the user will want to navigate to.